Move Map and all related class to Map subfolder
