+++ src/ts/Map/Cell.ts
@@ -1,0 +1,22 @@
+import { SimpleBox } from "../Libraries/Boxes";
+import { Brick } from "../Bricks";
+
+
+export class Cell extends SimpleBox {
+    brick: Brick;
+
+    constructor({ x, y, size, brick }: {
+        x: number;
+        y: number;
+        size: number;
+        brick: Brick;
+    }) {
+        super();
+
+        this.x = x;
+        this.y = y;
+        this.width = size;
+        this.height = size;
+        this.brick = brick;
+    }
+}
