+++ src/ts/Map/Map.ts
@@ -1,256 +1,22 @@
-import { Game } from "./Game";
+import { Game } from "../Game";
 
-import { SimpleBox } from "./Libraries/Boxes";
-import { AudioLoader } from "./Libraries/Loaders";
-import { Direction } from "./Libraries/Direction";
-import { Random } from "./Libraries/Random";
+import { AudioLoader } from "../Libraries/Loaders";
+import { Direction } from "../Libraries/Direction";
+import { Random } from "../Libraries/Random";
 
 import {
-  Enemy,
   Octorok,
   BlueOctorok,
   Moblin,
   BlueMoblin,
   Tektite,
   BlueTektite
-} from "./Enemies";
-import { Brick } from "./Bricks";
-import { HeartReceptacle, Item, Sword as SwordItem } from "./Items";
+} from "../Enemies";
 
-export class Cell extends SimpleBox {
-  brick: Brick;
-
-  constructor({ x, y, size, brick }: {
-    x: number;
-    y: number;
-    size: number;
-    brick: Brick;
-  }) {
-    super();
-
-    this.x = x;
-    this.y = y;
-    this.width = size;
-    this.height = size;
-    this.brick = brick;
-  }
-}
-
-export class Passage extends SimpleBox {
-  Game: Game;
-  Scene: Scene;
-
-  targetWorldIndex: number;
-  targetSceneC: number;
-  targetSceneR: number;
-
-  constructor({ game, scene, c, r, targetWorldIndex, targetSceneC, targetSceneR }: {
-    game: Game;
-    scene: Scene;
-    c: number;
-    r: number;
-    targetWorldIndex: number;
-    targetSceneC: number;
-    targetSceneR: number;
-  }) {
-    super();
-
-    this.Game = game;
-    this.Scene = scene;
-
-    this.x = c * this.Scene.cellSize;
-    this.y = r * this.Scene.cellSize;
-
-    this.width = this.Scene.cellSize;
-    this.height = this.Scene.cellSize;
-
-    this.targetWorldIndex = targetWorldIndex;
-    this.targetSceneC = targetSceneC;
-    this.targetSceneR = targetSceneR;
-  }
-}
-
-export class Scene {
-  Game: Game;
-  World: World;
-
-  cells: Cell[][] = [];
-
-  x: number;
-  y: number;
-
-  // Coordinates of the scene in the world
-  c: number;
-  r: number;
-
-  nbRow: number;
-  nbCol: number;
-  cellSize: number;
-
-  enemies: Enemy[];
-  passages: Passage[];
-  permanentItems: Item[];
-
-  music: HTMLAudioElement;
-
-  defaultBrick: Brick;
-  defaultWallBrick: Brick;
-
-  backgroundColor: string;
-
-  constructor({ game, world, c, r, music, defaultBrick, backgroundColor }: {
-    game: Game;
-    world: World;
-    c: number;
-    r: number;
-    music: HTMLAudioElement;
-    defaultBrick: Brick;
-    backgroundColor: string;
-  }) {
-    this.Game = game;
-    this.World = world;
-
-    this.nbRow = 11;
-    this.nbCol = 16;
-    this.cellSize = 64;
-
-    this.enemies = [];
-    this.passages = [];
-    this.permanentItems = [];
-
-    this.x = 0;
-    this.y = 0;
-
-    this.c = c;
-    this.r = r;
-
-    this.music = music;
-
-    this.backgroundColor = backgroundColor;
-
-    for (let c = 0; c < this.nbCol; c++) {
-      this.cells[c] = [];
-      for (let r = 0; r < this.nbRow; r++) {
-        this.cells[c][r] = new Cell({
-          x: this.cellSize * c,
-          y: this.cellSize * r,
-          size: this.cellSize,
-          brick: defaultBrick
-        });
-      }
-    }
-  }
-
-  public get width(): number {
-    return this.cellSize * this.nbCol;
-  }
-
-  public get height(): number {
-    return this.cellSize * this.nbRow;
-  }
-
-  get hasEnemies(): boolean {
-    return this.enemies.length > 0;
-  }
-
-  get hasPermanentItems(): boolean {
-    return this.permanentItems.length > 0;
-  }
-
-  getCell(col: number, row: number): Cell {
-    return this.cells[col][row];
-  }
-
-  loadBricks(bricks: any[][]): void {
-    bricks.forEach((row, r) => {
-      row.forEach((brickName, c) => {
-        this.cells[c][r].brick = this.Game.BrickCollection.get(brickName);
-      });
-    });
-  }
-
-  drawImage({ sprite, x, y, width, height }: {
-    sprite: HTMLImageElement;
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  }): void {
-    this.Game.Viewport.drawImage({ sprite, x: x + this.x, y: y + this.y, width, height });
-  }
-
-  fillRect({ x, y, width, height, color }: {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-    color: string;
-  }): void {
-    this.Game.Viewport.fillRect({ x: x + this.x, y: y + this.y, width, height, color });
-  }
-
-  upperEdgeCollision(): void {
-    this.Game.Viewport.slideScene(Direction.Up);
-  }
-
-  rightEdgeCollision(): void {
-    this.Game.Viewport.slideScene(Direction.Right);
-  }
-
-  bottomEdgeCollision(): void {
-    this.Game.Viewport.slideScene(Direction.Down);
-  }
-
-  leftEdgeCollision(): void {
-    this.Game.Viewport.slideScene(Direction.Left);
-  }
-}
-
-export class World {
-  Game: Game;
-
-  scenes: Scene[][] = [];
-
-  nbCol: number;
-  nbRow: number;
-
-  constructor({ game, nbCol, nbRow, defaultMusic, defaultBrick, defaultBackgroundColor }: {
-    game: Game;
-    nbCol: number;
-    nbRow: number;
-    defaultMusic: HTMLAudioElement;
-    defaultBrick: Brick;
-    defaultBackgroundColor: string;
-  }) {
-    this.Game = game;
-
-    this.nbCol = nbCol;
-    this.nbRow = nbRow;
-
-    for (let c = 0; c < this.nbCol; c++) {
-      this.scenes[c] = [];
-      for (let r = 0; r < this.nbRow; r++) {
-        this.scenes[c][r] = new Scene({
-          game: this.Game,
-          world: this,
-          c,
-          r,
-          music: defaultMusic,
-          defaultBrick,
-          backgroundColor: defaultBackgroundColor
-        });
-      }
-    }
-  }
-
-  loopScenes(callback: Function): void {
-    this.scenes.forEach((col) => {
-      col.forEach((scene) => {
-        callback(scene);
-      });
-    });
-  }
-}
+import { HeartReceptacle, Sword as SwordItem } from "../Items";
+import { Passage } from "./Passage";
+import { Scene } from "./Scene";
+import { World } from "./World";
 
 export class Map {
   Game: Game;
